Extract people loading
