+++ lib/data_loader/ukraine/people.rb
@@ -1,10 +1,63 @@
+require "open-uri"
+require "json"
+
 module DataLoader
   module Ukraine
     class People
       URL = ENV["DEBUG_URL"] || "https://raw.githubusercontent.com/everypolitician/everypolitician-data/master/data/Ukraine/Verkhovna_Rada/ep-popolo-v1.0.json"
 
       def self.load!
-        DataLoader::Ukraine::Popolo.load!(URL)
+        Rails.logger.info "Loading Ukraine vote events from #{URL}..."
+        data = JSON.parse(open(URL).read)
+
+        if people = data["persons"]
+          organizations = data["organizations"]
+          areas = data["areas"]
+          events = data["events"]
+
+          Rails.logger.info "Loading #{people.count} people..."
+          people.each do |p|
+            person = Person.find_or_initialize_by(id: extract_rada_id_from_person(p))
+            person.small_image_url = p["image"]
+            person.large_image_url = p["image"]
+            person.save!
+          end
+
+          members = data["memberships"]
+          Rails.logger.info "Loading #{members.count} memberships..."
+          members.each do |m|
+            raise "Person not found: #{m["person_id"]}" unless person = people.find { |p| p["id"] == m["person_id"] }
+            raise "Party not found: #{m["on_behalf_of_id"]}" unless party = organizations.find { |o| o["id"] == m["on_behalf_of_id"] }
+            raise "Area not found: #{m["area_id"]}" unless area = areas.find { |a| a["id"] == m["area_id"] }
+            raise "Legislative period not found: #{m["legislative_period_id"]}" unless legislative_period = events.find { |e| e["id"] == m["legislative_period_id"] }
+            person["rada_id"] = extract_rada_id_from_person(person)
+
+            # Default to the start of the legislative period if there no specific one set for this membership
+            start_date = m["start_date"] || legislative_period["start_date"]
+
+            member = Member.find_or_initialize_by(person_id: person["rada_id"], entered_house: start_date)
+            member.gid = m["person_id"]
+            member.source_gid = person["rada_id"]
+            member.first_name = person["given_name"]
+            member.last_name = person["family_name"]
+            member.title = ""
+            member.constituency = area["name"]
+            member.party = party["name"]
+            # TODO: Remove hardcoded house
+            member.house = "rada"
+            member.entered_house = start_date
+            member.left_house = m["end_date"] if m["end_date"]
+            member.person_id = person["rada_id"]
+            member.save!
+          end
+        else
+          raise "No loadable data found"
+        end
+      end
+
+      # TODO: This shouldn't be a class method - move it somewhere more sensible
+      def self.extract_rada_id_from_person(person)
+        person["identifiers"].find { |i| i["scheme"] == "rada" }["identifier"]
       end
     end
   end
